feat: add norminette to libs
