+++ .github/workflows/ci.yml
@@ -15,7 +15,7 @@ jobs:
         run: |
           pip install -U pip
           pip install norminette
-          norminette $(find src/ -type f \( -iname "*.c" ! -iname "wip_*" \)) includes/
+          norminette $(find src/ -type f \( -iname "*.c" ! -iname "wip_*" \)) includes/ libs/builtins/ libs/libft libs/get_next_line
   build:
     runs-on: ubuntu-latest
     steps:
