+++ Makefile
@@ -6,7 +6,7 @@
 #    By: vwildner <vwildner@student.42sp.org.br>    +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/01/19 22:38:45 by itaureli          #+#    #+#              #
-#    Updated: 2022/06/01 20:57:57 by vwildner         ###   ########.fr        #
+#    Updated: 2022/06/01 22:36:21 by vwildner         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -56,7 +56,7 @@ TESTS = $(addprefix $(TEST_PATH)/,$(TEST_FILES))
 
 BUILTINS_NAME = libbuiltins.a
 
-BUILTINS_SOURCES_FILES		= builtins.c echo.c dispatcher.c env.c export.c builtins_export.c utils.c
+BUILTINS_SOURCES_FILES		= builtins.c echo.c dispatcher.c env.c export.c builtins_export.c builtin_utils.c builtins_cd.c builtin_utils2.c
 
 BUILTINS_SOURCES_PATH = ./libs/builtins
