+++ libs/builtins/builtin_utils.c
@@ -6,7 +6,7 @@
 /*   By: vwildner <vwildner@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/05 22:27:35 by vwildner          #+#    #+#             */
-/*   Updated: 2022/05/28 21:49:22 by vwildner         ###   ########.fr       */
+/*   Updated: 2022/06/01 22:39:13 by vwildner         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,50 +56,6 @@ t_command	*init_builtins(char *envp[])
 	return (cmd);
 }
 
-static int are_equal(char *one, char *other)
-{
-	return (ft_strncmp(one, other, ft_strlen(other)));
-}
-
-t_list	*lst_find(t_list **list, char *key)
-{
-	t_list	*tmp;
-
-	tmp = *list;
-	while (tmp != NULL)
-	{
-		if (are_equal(tmp->key, key) == 0)
-			return (tmp);
-		tmp = tmp->next;
-	}
-	return (NULL);
-}
-
-int	lst_del_node(t_list **list, char *key)
-{
-	t_list	*tmp;
-	t_list	*prev;
-
-	tmp = *list;
-	prev = NULL;
-	if (tmp != NULL && are_equal(tmp->key, key) == 0)
-	{
-		*list = tmp->next;
-		free(tmp);
-		return (1);
-	}
-	while (tmp != NULL && are_equal(tmp->key, key) != 0)
-	{
-		prev = tmp;
-		tmp = tmp->next;
-	}
-	if (tmp == NULL)
-		return (0);
-	prev->next = tmp->next;
-	free(tmp);
-	return (1);
-}
-
 char	*ms_getenv(t_list *envp[], char *key)
 {
 	t_list			*tmp;
@@ -123,17 +79,4 @@ int	read_file(char *filename)
 	dup2(fd, 0);
 	close(fd);
 	return (fd);
-}
-
-void	print_err_msg(char *command, char *msg)
-{
-	char	*err_msg;
-
-	err_msg = ft_strdup("minishell: ");
-	err_msg = ft_strjoin(err_msg, command);
-	err_msg = ft_strjoin(err_msg, ": ");
-	err_msg = ft_strjoin(err_msg, msg);
-	err_msg = ft_strjoin(err_msg, "\n");
-	write(STDERR_FILENO, err_msg, ft_strlen(err_msg));
-	free(err_msg);
 }
