+++ libs/builtins/builtins.c
@@ -1,6 +1,18 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   builtins.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: vwildner <vwildner@student.42sp.org.br>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/06/01 22:23:50 by vwildner          #+#    #+#             */
+/*   Updated: 2022/06/01 22:25:56 by vwildner         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "../../includes/builtins.h"
 
-int builtins_echo(t_command *cmd)
+int	builtins_echo(t_command *cmd)
 {
 	if (cmd->argc <= 1)
 	{
@@ -11,37 +23,9 @@ int builtins_echo(t_command *cmd)
 	return (0);
 }
 
-int builtins_env(t_command *cmd)
+int	builtins_env(t_command *cmd)
 {
 	env(cmd->envp, cmd->fd);
-	return (0);
-}
-
-int	builtins_cd(t_command *cmd)
-{
-	char	buf[1025];
-	int		status;
-
-	status = 0;
-	getcwd(buf, 1024);
-	if (cmd->argc == 1)
-	{
-		cd_home(cmd->envp);
-		return (set_oldpwd(cmd->envp, buf));
-	}
-	if (cmd->argc > 2)
-	{
-		print_err_msg(cmd->argv[0], "too many arguments");
-		return (127);
-	}
-	if (cmd->argv[1][0] == '-')
-	{
-		chdir(ms_getenv(cmd->envp, "OLDPWD"));
-		return (set_oldpwd(cmd->envp, buf));
-	}
-	status = cd(cmd->argv, cmd->envp);
-	if (status == 0)
-		return (set_oldpwd(cmd->envp, buf));
 	return (0);
 }
