+++ libs/builtins/cd.c
@@ -6,7 +6,7 @@
 /*   By: vwildner <vwildner@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/11 20:31:33 by itaureli          #+#    #+#             */
-/*   Updated: 2022/05/28 21:49:47 by vwildner         ###   ########.fr       */
+/*   Updated: 2022/06/01 22:43:07 by vwildner         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,24 +27,24 @@ int	set_oldpwd(t_list *envp[], char *buf)
 	return (0);
 }
 
-int cd_home(t_list *envp[])
+int	cd_home(t_list *envp[])
 {
-	char *home;
+	char	*home;
 
-	home = NULL;
-	if ((home = ms_getenv(envp, "HOME")) == NULL)
+	home = ms_getenv(envp, "HOME");
+	if (home == NULL)
 		return (1);
 	if (chdir(home) == -1)
 		return (1);
 	return (0);
 }
 
-static void print_cd_error(char *argv[])
+static void	print_cd_error(char *argv[])
 {
-		write(STDERR_FILENO, "minishell: cd: ", 10);
-		write(STDERR_FILENO, argv[1], ft_strlen(argv[1]));
-		write(STDERR_FILENO, ": No such file or directory", 27);
-		write(STDERR_FILENO, "\n", 1);
+	write(STDERR_FILENO, "minishell: cd: ", 10);
+	write(STDERR_FILENO, argv[1], ft_strlen(argv[1]));
+	write(STDERR_FILENO, ": No such file or directory", 27);
+	write(STDERR_FILENO, "\n", 1);
 }
 
 // TO-DO: refactor cd
