+++ libs/builtins/echo.c
@@ -6,7 +6,7 @@
 /*   By: vwildner <vwildner@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/20 04:54:03 by itaureli          #+#    #+#             */
-/*   Updated: 2022/06/01 22:22:54 by vwildner         ###   ########.fr       */
+/*   Updated: 2022/06/01 22:42:03 by vwildner         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,8 +21,8 @@ static int	check_flag_n(char *args)
 
 int	echo(char **args, int fd)
 {
-	int i;
-	int has_n;
+	int	i;
+	int	has_n;
 
 	i = 1;
 	has_n = check_flag_n(args[i]);
